Split crate::* on crate::atom and crate::space modules
Move Atom API from crate to crate::atom. Move matching functions
from crate::matcher to crate::atom::matcher. Move grounding space
implementation from crate to crate::space::grounding.
