+++ lib/src/arithmetics.rs
@@ -78,6 +78,9 @@ mod tests {
 
     use super::*;
     use crate::interpreter::*;
+    use crate::atom::matcher::Bindings;
+    use crate::space::grounding::GroundingSpace;
+
     // Aliases to have a shorter notation
     fn G<T: GroundedAtom>(value: T) -> Atom { Atom::gnd(value) }
 
@@ -114,8 +117,7 @@ mod tests {
         // (= (fac n) (* n (fac (- n 1))))
         space.add(expr!("=", ("fac", n), ({MUL}, n, ("fac", ({SUB}, n, {1})))));
 
-        let expected: Bindings = Bindings([(VariableAtom::from("X"),
-            expr!({MUL}, {3}, ("fac", ({SUB}, {3}, {1}))))].iter().cloned().collect());
+        let expected = bind!{X: expr!({MUL}, {3}, ("fac", ({SUB}, {3}, {1})))};
         assert_eq!(space.query(&expr!("=", ("fac", {3}), X)), vec![expected]);
     }
