+++ lib/src/atom/mod.rs
@@ -1,0 +1,319 @@
+// Macros to simplify expression writing
+
+#[macro_export]
+macro_rules! expr {
+    () => { Atom::expr(&[]) };
+    ($x:ident) => { Atom::var(stringify!($x)) };
+    ($x:literal) => { Atom::sym($x) };
+    ({$x:tt}) => { Atom::gnd($x) };
+    (($($x:tt),*)) => { Atom::expr(&[ $( expr!($x) , )* ]) };
+    ($($x:tt),*) => { Atom::expr(&[ $( expr!($x) , )* ]) };
+}
+
+pub mod matcher;
+
+use std::fmt::{Display, Debug};
+
+// Symbol atom
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct SymbolAtom {
+    name: String,
+}
+
+impl From<String> for SymbolAtom {
+    fn from(name: String) -> Self {
+        SymbolAtom{ name }
+    }
+}
+
+impl From<&str> for SymbolAtom {
+    fn from(name: &str) -> Self {
+        SymbolAtom::from(name.to_string())
+    }
+}
+
+impl SymbolAtom {
+    pub fn name(&self) -> &str {
+        self.name.as_str()
+    }
+}
+
+impl Display for SymbolAtom {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.name)
+    }
+}
+
+// Expression atom
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct ExpressionAtom {
+    children: Vec<Atom>,
+}
+
+impl ExpressionAtom {
+    pub fn is_plain(&self) -> bool {
+        self.children.iter().all(|atom| ! matches!(atom, Atom::Expression(_)))
+    }
+
+    pub fn children(&self) -> &Vec<Atom> {
+        &self.children
+    }
+
+    pub fn children_mut(&mut self) -> &mut Vec<Atom> {
+        &mut self.children
+    }
+}
+
+impl From<Vec<Atom>> for ExpressionAtom {
+    fn from(children: Vec<Atom>) -> Self {
+        ExpressionAtom{ children }
+    }
+}
+
+impl From<&[Atom]> for ExpressionAtom {
+    fn from(children: &[Atom]) -> Self {
+        ExpressionAtom::from(children.to_vec())
+    }
+}
+
+impl Display for ExpressionAtom {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "(")
+            .and_then(|_| self.children.iter().take(1).fold(Ok(()),
+                |res, atom| res.and_then(|_| write!(f, "{}", atom))))
+            .and_then(|_| self.children.iter().skip(1).fold(Ok(()),
+                |res, atom| res.and_then(|_| write!(f, " {}", atom))))
+            .and_then(|_| write!(f, ")"))
+    }
+}
+
+// Variable atom
+
+#[derive(Debug, Clone, Hash, PartialEq, Eq)]
+pub struct VariableAtom {
+    name: String,
+}
+
+impl From<String> for VariableAtom {
+    fn from(name: String) -> Self {
+        VariableAtom{ name }
+    }
+}
+
+impl From<&str> for VariableAtom {
+    fn from(name: &str) -> Self {
+        VariableAtom::from(name.to_string())
+    }
+}
+
+impl VariableAtom {
+    pub fn name(&self) -> &str {
+        self.name.as_str()
+    }
+}
+
+impl Display for VariableAtom {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "${}", self.name)
+    }
+}
+
+// Grounded atom
+
+pub trait GroundedAtom : Debug + mopa::Any {
+    fn execute(&self, _args: &mut Vec<Atom>) -> Result<Vec<Atom>, String> {
+        Err(format!("{:?} is not executable", self))
+    }
+    fn eq_gnd(&self, other: &dyn GroundedAtom) -> bool;
+    fn clone_gnd(&self) -> Box<dyn GroundedAtom>;
+}
+
+mopafy!(GroundedAtom);
+
+// FIXME: one cannot implement custom execute() for the type which
+// have derived Clone, PartialEq and Debug at the same time because
+// it automatically have GroundedAtom implemented from the impl below.
+// This is a Rust restriction: there is not method overriding and
+// trait can be implemented only once for each type.
+
+// GroundedAtom implementation for all "regular" types
+// to allow using them as GroundedAtoms
+// 'static is required because mopa::Any requires it
+impl<T: 'static + Clone + PartialEq + Debug> GroundedAtom for T {
+    fn eq_gnd(&self, other: &dyn GroundedAtom) -> bool {
+        match other.downcast_ref::<T>() {
+            Some(o) => self.eq(o),
+            None => false,
+        }
+    }
+
+    fn clone_gnd(&self) -> Box<dyn GroundedAtom> {
+        Box::new(self.clone())
+    }
+}
+
+// Atom enum
+
+pub enum Atom {
+    Symbol(SymbolAtom),
+    Expression(ExpressionAtom),
+    Variable(VariableAtom),
+    // We need using Box here because:
+    // - we cannot use GroundedAtom because trait size is not known at compile time
+    // - reference to trait does not allow heap allocated values
+    // - other smart pointers like Rc doesn't allow choosing whether value should
+    //   be copied or shared between two atoms when clone() is called
+    Grounded(Box<dyn GroundedAtom>),
+}
+
+impl Atom {
+    pub fn sym(name: &str) -> Self {
+        Self::Symbol(SymbolAtom::from(name))
+    }
+
+    pub fn expr(children: &[Atom]) -> Self {
+        Self::Expression(ExpressionAtom::from(children))
+    }
+
+    pub fn var(name: &str) -> Self {
+        Self::Variable(VariableAtom::from(name))
+    }
+
+    pub fn gnd<T: GroundedAtom>(gnd: T) -> Atom {
+        Self::Grounded(Box::new(gnd))
+    }
+
+    pub fn as_gnd<T: GroundedAtom>(&self) -> Option<&T> {
+        match self {
+            Atom::Grounded(gnd) => gnd.downcast_ref::<T>(),
+            _ => None,
+        }
+    }
+
+    pub fn as_gnd_mut<T: GroundedAtom>(&mut self) -> Option<&mut T> {
+        match self {
+            Atom::Grounded(gnd) => gnd.downcast_mut::<T>(),
+            _ => None,
+        }
+    }
+}
+
+impl PartialEq for Atom {
+    fn eq(&self, other: &Self) -> bool {
+        match (self, other) {
+            (Atom::Symbol(sym1), Atom::Symbol(sym2)) => sym1 == sym2,
+            (Atom::Expression(expr1), Atom::Expression(expr2)) => expr1 == expr2,
+            (Atom::Variable(var1), Atom::Variable(var2)) => var1 == var2,
+            (Atom::Grounded(gnd1), Atom::Grounded(gnd2)) => gnd1.eq_gnd(&**gnd2),
+            _ => false,
+        }
+    }
+}
+
+impl Eq for Atom {}
+
+impl Clone for Atom {
+    fn clone(&self) -> Self {
+        match self {
+            Atom::Symbol(sym) => Atom::Symbol(sym.clone()),
+            Atom::Expression(expr) => Atom::Expression(expr.clone()),
+            Atom::Variable(var) => Atom::Variable(var.clone()),
+            Atom::Grounded(gnd) => Atom::Grounded((*gnd).clone_gnd()),
+        }
+    }
+}
+
+impl Display for Atom {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Atom::Symbol(sym) => Display::fmt(sym, f),
+            Atom::Expression(expr) => Display::fmt(expr, f),
+            Atom::Variable(var) => Display::fmt(var, f),
+            Atom::Grounded(gnd) => Debug::fmt(gnd, f),
+        }
+    }
+}
+
+impl Debug for Atom {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        Display::fmt(self, f)
+    }
+}
+
+#[cfg(test)]
+mod test {
+    #![allow(non_snake_case)]
+
+    use super::*;
+    use std::collections::HashMap;
+
+    // Aliases to have a shorter notation
+    fn S(name: &str) -> Atom { Atom::sym(name) }
+    fn E(children: &[Atom]) -> Atom { Atom::expr(children) }
+    fn V(name: &str) -> Atom { Atom::var(name) }
+    fn G<T: GroundedAtom>(gnd: T) -> Atom { Atom::gnd(gnd) }
+
+    #[test]
+    fn test_expr_symbol() {
+        assert_eq!(expr!("="), S("="));
+        assert_eq!(expr!("1"), S("1"));
+        assert_eq!(expr!("*"), S("*"));
+        assert_eq!(expr!("foo"), S("foo"));
+    }
+
+    #[test]
+    fn test_expr_variable() {
+        assert_eq!(expr!(n), V("n"));
+        assert_eq!(expr!(self), V("self"));
+    }
+
+    #[test]
+    fn test_expr_expression() {
+        assert_eq!(expr!("=", ("fact", n), ("*", n, ("-", n, "1"))), 
+            E(&[S("="), E(&[S("fact"), V("n")]),
+            E(&[ S("*"), V("n"), E(&[ S("-"), V("n"), S("1") ]) ]) ]));
+        assert_eq!(expr!("=", n, {[1, 2, 3]}), E(&[S("="), V("n"), G([1, 2, 3])]));
+        assert_eq!(expr!("=", {6}, ("fact", n)), E(&[S("="), G(6), E(&[S("fact"), V("n")])]));
+    }
+
+    #[test]
+    fn test_grounded() {
+        assert_eq!(Atom::gnd(3), Atom::Grounded(Box::new(3)));
+        assert_eq!(G(42).as_gnd::<i32>().unwrap(), &42);
+        assert_eq!(G("Data string"), Atom::Grounded(Box::new("Data string")));
+        assert_eq!(G(vec![1, 2, 3]), Atom::Grounded(Box::new(vec![1, 2, 3])));
+        assert_eq!(G([42, -42]).as_gnd::<[i32; 2]>().unwrap(), &[42, -42]);
+        assert_eq!(G((-42, "42")).as_gnd::<(i32, &str)>().unwrap(), &(-42, "42"));
+        assert_eq!(G(HashMap::from([("q", 0), ("a", 42),])),
+        Atom::Grounded(Box::new(HashMap::from([("q", 0), ("a", 42),]))));
+    }
+
+    #[test]
+    fn test_display_symbol() {
+        assert_eq!(format!("{}", Atom::sym("test")), "test");
+    }
+
+    #[test]
+    fn test_display_variable() {
+        assert_eq!(format!("{}", Atom::var("x")), "$x");
+    }
+
+    #[test]
+    fn test_display_expression() {
+        assert_eq!(format!("{}", expr!("=", ("fact", n), ("*", n, ("-", n, "1")))),
+        "(= (fact $n) (* $n (- $n 1)))");
+        assert_eq!(format!("{}", expr!()), "()");
+    }
+
+    #[test]
+    fn test_display_grounded() {
+        assert_eq!(format!("{}", Atom::gnd(42)), "42");
+        assert_eq!(format!("{}", Atom::gnd([1, 2, 3])), "[1, 2, 3]");
+        assert_eq!(
+            format!("{}", Atom::gnd(HashMap::from([("hello", "world")]))),
+            "{\"hello\": \"world\"}");
+    }
+
+}
