+++ lib/src/text.rs
@@ -1,4 +1,5 @@
 use crate::*;
+use crate::space::grounding::GroundingSpace;
 
 use std::io::Read;
 use regex::Regex;
